Improve the PHP code
The optimizer uses type informations, so adding types should result in slightly better execution time. Also, the code will be more robust and readable.

Using dynamic properties is slow, and even deprecated as of PHP 8.2. So changed to declare explicitly. This also improves robustness a bit as it aids static analysis in IDEs and editors.

Usually we don't use closing PHP tags outside of templating uses. This is because unintentionally added whitespaces after the closing tag are added to the output.